requirements.txt
RPi.GPIO==0.7.1
smbus2==0.5.0
